Move par → parallel
